Extract middleman helpers into a module
